requirements.txt
Django
mysqlclient
djangorestframework
configparser
django-cors-headers